+++ CHANGELOG.md
@@ -31,6 +31,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - creation of advancement tab when an advancement NotifyIO is sent
 - `burn` event throwing unexpected error when omitting duration
 - added missing firework meta to string representation of QuestItem
+- required flags in `QuestItem` were not relevant in equals check
 ### Security
 
 ## [2.0.1] - 2024-03-24